requirements.txt
absl-py==0.7.0
astor==0.7.1
bunch==1.0.1
certifi==2018.11.29
cycler==0.10.0
gast==0.2.2
grpcio==1.12.1
h5py==2.7.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
mkl-fft==1.0.10
mkl-random==1.0.2
numpy==1.15.4
olefile==0.46
opencv-python==4.0.0.21
Pillow==5.4.1
protobuf==3.6.1
pyparsing==2.3.1
python-dateutil==2.8.0
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tqdm==4.31.1
Werkzeug==0.14.1
wincertstore==0.2
